initial implementation of the commitment strategy

+++ src/main/java/com/easyhelp/application/model/dto/requests/DonationRequestDTO.java
@@ -2,6 +2,7 @@ package com.easyhelp.application.model.dto.requests;
 
 import com.easyhelp.application.model.blood.BloodComponent;
 import com.easyhelp.application.model.dto.BaseDTO;
+import com.easyhelp.application.model.requests.DonationRequest;
 import com.easyhelp.application.model.requests.RequestUrgency;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -12,7 +13,7 @@ public class DonationRequestDTO extends BaseDTO {
 
     private Long doctorId;
     private Long patientId;
-    private Integer quantity;
+    private Double quantity;
     private RequestUrgency urgency;
     private BloodComponent bloodComponent;
 }
